requirements.txt
openai-whisper
pyttsx3
opencv-python
sounddevice
numpy
scipy
ollama
Pillow
openwakeword
mediapipe
